requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
langchain
faiss-cpu
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
PyPDF2
python-multipart          